Add basic scripts for running in dev and prod
Add scripts and supporting libraries for running the application
in dev and prod environments, handling ES2015 transpilation and
serving assets.

For now, in development mode, the application can only be
accessed from the 'localhost:8080/dev' URL. Could not find an
alternative that would work with Webpack Dev Server for now.